Update for start pulse for DHT sensors
Datasheet for DHT22 / AM2302 shows 'min 1ms' rather than 20ms required for DHT11.  Also changing from pulling the pin high to releasing it to INPUT_PULLUP per thread on issue #5619  This is not a complete fix for that issue as it appears to be device 'brand' related but it does improve reliability of DHT22's in testing on WemosD1 and ESP01 devices I had to hand.  DHT11 tested as well.